Move parameter class in separate include file
